wip: add assignment functionality to backend
