+++ backend/src/routes/courseRouter.ts
@@ -1,0 +1,19 @@
+import express from "express";
+import {
+  getAllCourses,
+  getAssignments,
+} from "../controllers/courseController.js";
+
+const router = express.Router();
+
+/**
+ * @route - GET courses/
+ */
+router.get("/", getAllCourses);
+
+/**
+ * @route - GET courses/:course_id/assignments
+ */
+router.get("/:courseId/assignments", getAssignments);
+
+export default router;
